Define channel island (purpose of visit question)
Addition of the channel island and isle of man question, the intention
here is to provide intending transit visitors to the channel isles a similar
question (with same options except for transit).

The logic that will follow the selection of any of the options will behave
the same with its equivalent from the purpose of visit question.