Implement transit rules for non visa nationals
Here I have added rules to cater for transit journey enquiries for non visa nationals

This groups is capture in with the COUNTRY_GROUP_NON_VISA_NATIONAL in Calculators::UkVisaCalculator.

As requested by Content designers, I have added a detour for Non Visa Nationals when they have chosen the transit option,
and also added another path for the aforementioned group when republic of Ireland is chosen.