Regenerate test artefacts for Bulgaria (Marriage Abroad)
This commit shows the affected/changes to regression test artefacts as a result of the
changes to the for all opposite sex outcomes where user is resident in the Bulgaria.

I have reviewed the changes to the regression test artefacts and it is as expected.

The changes affects only the ceremony country route of the Bulgarian marriage abroad smart answer

Already existing unit, regression and integration tests passed, giving assurance
that the aforementioned commit/changes have had no effect other parts the code and only affect the
ceremony country route of the Bulgarian marriage abroad smart answer