Update transit to Ireland for Non-Visa-National & Oversea Territory
This commit makes logic and content changes to the transit route for Overseas Territory and
Non-Visa-National countries/passport holders with the intention of visiting the UK on transit to
the Republic of Ireland (a Common Travel Area i.e CTA).

In this commit the focus is Scenario B in the doc[1], which is for the Republic of Ireland transit option.

In some areas I didn't need to make changes for Non-Visa-National as it was already as required.

I have included added content to the no visa needed outcome, which will only show for the aforementioned groups.

I have updated the tests and made changes to the content as recommended in the google doc [1].

[1] https://docs.google.com/document/d/12PnaBt8B797xibETodkUdNTaBTMqN5yfWbvt275XAUU/edit?ts=57bc67cd#