Regenerate regression tests artefacts (Check UK Visa)
This commit shows the affected/changes to regression test artefacts as a result
of the changes from (immediate) previous commits.

I have reviewed the changes to the regression test artefacts and it is as expected.

The changes affects only the transit route of the check uk visa smart answer

Already existing unit, regression and integration tests passed, giving assurance
that the aforementioned changes have had no (interfering) effect to the transit
route or flow or other routes within check uk visa or to other smart answers.