Generate regression tests artefacts (SEC)
This commit shows the affected/changes to regression test artefacts as a result
of the changes (i.e addition CO2 emission options) from previous commits.

I have reviewed the changes to the regression test artefacts and it is as
expected.

This commit creates regression artefacts that are inline with the new
introduced by the addition of CO2 emission options, but still contain content
that will change in the next set of commits.

At this point existing unit, regression and integration tests passed, giving
assurance that the aforementioned changes have had no (interfering) affect to
other smart answers.