TEMP: Regenerate regression tests artefacts (Check UK Visa)
This commit shows the affected/changes to regression test artefacts as a result
of the changes from (immediate) previous commits.

I have reviewed the changes to the regression test artefacts and it is as expected.

The changes affect the electronic waiver countries (kuwait and oman) and datv
countries on transit to CTA.

Affected outcomes:

https://smart-answers-pr-2594.herokuapp.com/check-uk-visa/y/afghanistan/transit/channel_islands_or_isle_of_man
https://smart-answers-pr-2594.herokuapp.com/check-uk-visa/y/kuwait/transit/somewhere_else/yes
https://smart-answers-pr-2594.herokuapp.com/check-uk-visa/y/oman/transit/somewhere_else/yes
https://smart-answers-pr-2594.herokuapp.com/check-uk-visa/y/stateless-or-refugee/transit/somewhere_else/yes

The changes affects only the transit route of the check uk visa smart answer

Already existing unit, regression and integration tests passed, giving assurance
that the aforementioned changes have had no (interfering) effect to the transit
route or flow or other routes within check uk visa or to other smart answers.